+++ packages/i18n/src/i18n-manager/translations-manager/translations-loaders/routeCreateTranslationLoader.ts
@@ -1,4 +1,4 @@
-import { CustomMapping } from 'generaltranslation/types';
+import type { CustomMapping } from 'generaltranslation/types';
 import { TranslationsLoader } from './types';
 import { LoadTranslationsType } from '../../utils/getLoadTranslationsType';
 import logger from '../../../logs/logger';