+++ packages/react/CHANGELOG.md
@@ -1,5 +1,25 @@
 # gt-react
 
+## 10.19.7
+
+### Patch Changes
+
+- Updated dependencies [[`ce0933a`](https://github.com/generaltranslation/gt/commit/ce0933ab102d34a0c38634f7c2b0d634c9a620a8)]:
+  - generaltranslation@8.2.9
+  - gt-i18n@0.8.8
+  - @generaltranslation/react-core@1.8.9
+  - @generaltranslation/supported-locales@2.0.67
+
+## 10.19.6
+
+### Patch Changes
+
+- Updated dependencies [[`b12d57d`](https://github.com/generaltranslation/gt/commit/b12d57dab1d5cb1f602c5ac24a702b48cda7f11e)]:
+  - generaltranslation@8.2.8
+  - gt-i18n@0.8.7
+  - @generaltranslation/react-core@1.8.8
+  - @generaltranslation/supported-locales@2.0.66
+
 ## 10.19.5
 
 ### Patch Changes